+++ ui/resolution.go
@@ -9,9 +9,9 @@ import (
 	raylib "github.com/gen2brain/raylib-go/raylib"
 )
 
-func GetResolution() (int, int) {
-	width := raylib.GetScreenWidth()
-	height := raylib.GetScreenHeight()
+func GetResolution() (int32, int32) {
+	width := int32(raylib.GetScreenWidth())
+	height := int32(raylib.GetScreenHeight())
 
 	if width == 0 || height == 0 {
 		width, height = getSysResolution()
@@ -20,7 +20,7 @@ func GetResolution() (int, int) {
 	return width, height
 }
 
-func getSysResolution() (int, int) {
+func getSysResolution() (int32, int32) {
 	switch runtime.GOOS {
 	case "darwin":
 		return getSysResolution_darwin()
@@ -36,7 +36,7 @@ func getSysResolution() (int, int) {
 	}
 }
 
-func getSysResolution_darwin() (int, int) {
+func getSysResolution_darwin() (int32, int32) {
 	cmd := exec.Command("system_profiler", "SPDisplaysDataType")
 	out, _ := cmd.Output()
 	for _, line := range strings.Split(string(out), "\n") {
@@ -45,13 +45,13 @@ func getSysResolution_darwin() (int, int) {
 			resSplit := strings.Split(res, " x ")
 			width, _ := strconv.Atoi(resSplit[0])
 			height, _ := strconv.Atoi(resSplit[1])
-			return width, height
+			return int32(width), int32(height)
 		}
 	}
 	return 0, 0
 }
 
-func getSysResolution_linux() (int, int) {
+func getSysResolution_linux() (int32, int32) {
 	cmd := exec.Command("xrandr")
 	if out, err := cmd.Output(); err == nil {
 		for _, line := range strings.Split(string(out), "\n") {
@@ -60,21 +60,21 @@ func getSysResolution_linux() (int, int) {
 				resSplit := strings.Split(res, "x")
 				width, _ := strconv.Atoi(resSplit[0])
 				height, _ := strconv.Atoi(resSplit[1])
-				return width, height
+				return int32(width), int32(height)
 			}
 		}
 	}
 	return 0, 0
 }
 
-func getSysResolution_windows() (int, int) {
+func getSysResolution_windows() (int32, int32) {
 	// XXX: untested!!
 	cmd := exec.Command("wmic", "desktopmonitor", "get", "screenwidth", "screenheight")
 	if res, err := cmd.Output(); err == nil {
 		data := strings.Split(string(res), " ")
 		width, _ := strconv.Atoi(data[0])
 		height, _ := strconv.Atoi(data[1])
-		return width, height
+		return int32(width), int32(height)
 	}
 	return 0, 0
 }
